pyproject.toml
[project]
name = "activity-logger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.0.0",
    "pyscreenshot>=3.1",
]
